Do not use host networking
Using host networking causes conflicts if there are already local ports
running on the port required by Microshift. We can avoid this problem by
running Microshift and the registry in the same container network namespace
and publishing the registry on a random host port (so that we're still able
to push images to it).